Before sending game states, server sends whether game has ended
This notifies client when the game has ended in a cleaner manner than was used before, which checked if the message that was supposed to be the game was long enough. If it wasn't, it had to be the message asking for the rematch which meant the game had ended.

Rework TUI client to work with new server